# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ACIM Cloud Architecture Success Cases Portal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Offering {
  id: number;
  name: string;
  durationHours: string;
  techPillar: string;
  businessBenefit: string;
  whatIsIt: string;
  whatDoesItSolve: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateOfferingBody {
  name: string;
  durationHours: string;
  techPillar: string;
  businessBenefit: string;
  whatIsIt: string;
  whatDoesItSolve: string;
}

export interface UpdateOfferingBody {
  name?: string;
  durationHours?: string;
  techPillar?: string;
  businessBenefit?: string;
  whatIsIt?: string;
  whatDoesItSolve?: string;
}

export interface UseCase {
  id: number;
  offeringId: number;
  companyIconUrl: string | null;
  projectName: string;
  industryType: string;
  description: string;
  previousState: string;
  newState: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateUseCaseBody {
  offeringId: number;
  companyIconUrl?: string | null;
  projectName: string;
  industryType: string;
  description: string;
  previousState: string;
  newState: string;
}

export interface UpdateUseCaseBody {
  offeringId?: number;
  companyIconUrl?: string | null;
  projectName?: string;
  industryType?: string;
  description?: string;
  previousState?: string;
  newState?: string;
}

export interface AzureAnalyzerKpis {
  governanceScore: number;
  wellArchitectedScore: number;
  reliabilityScore: number;
  securityScore: number;
  costOptimizationScore: number;
  operationalExcellenceScore: number;
  performanceScore: number;
  totalResources: number;
  managedResources: number;
  findings: number;
}

export interface AzureAuditorFindings {
  summary: string;
  criticalIssues: string[];
  recommendations: string[];
  costSavingsEstimate: number;
  costSavingsCurrency: string;
  complianceStatus: string;
  auditDate: string;
}

export interface SuccessCase {
  id: number;
  title: string;
  clientName: string;
  executiveSummary: string;
  azureAnalyzerKpis: AzureAnalyzerKpis;
  azureAuditorFindings: AzureAuditorFindings;
  discoveryMapUrl: string | null;
  industry: string | null;
  publishedAt: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateSuccessCaseBody {
  title: string;
  clientName: string;
  executiveSummary: string;
  azureAnalyzerKpis: AzureAnalyzerKpis;
  azureAuditorFindings: AzureAuditorFindings;
  discoveryMapUrl?: string | null;
  industry?: string | null;
  publishedAt?: string | null;
}

export interface UpdateSuccessCaseBody {
  title?: string;
  clientName?: string;
  executiveSummary?: string;
  azureAnalyzerKpis?: AzureAnalyzerKpis;
  azureAuditorFindings?: AzureAuditorFindings;
  discoveryMapUrl?: string | null;
  industry?: string | null;
  publishedAt?: string | null;
}

export type SuccessCasesStatsTopIndustriesItem = {
  industry: string;
  count: number;
};

export interface SuccessCasesStats {
  totalCases: number;
  totalClients: number;
  avgGovernanceScore: number;
  avgWellArchitectedScore: number;
  totalCostSavings: number;
  topIndustries: SuccessCasesStatsTopIndustriesItem[];
  recentCases: SuccessCase[];
}

export interface AdminLoginBody {
  username: string;
  password: string;
}

export interface AdminLoginResponse {
  token: string;
  username: string;
}

export interface AdminSession {
  username: string;
  authenticated: boolean;
}

export type ListOfferings200 = {
  offerings: Offering[];
};

export type ListUseCasesParams = {
  offeringId?: number;
  industryType?: string;
  search?: string;
};

export type ListUseCases200 = {
  useCases: UseCase[];
};

export type ListSuccessCasesParams = {
  search?: string;
  limit?: number;
  offset?: number;
};

export type ListSuccessCases200 = {
  cases: SuccessCase[];
  total: number;
};